14. Longest Common Prefix
